refactor(*): Extract path utils from common utils module
